requirements.txt
asgiref==3.8.1
Django==5.0.4
image==1.5.33
pillow==10.3.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
